Move projects methods from MiningFramework to Project
