+++ src/main/app/MiningFramework.groovy
@@ -1,13 +1,9 @@
 package main.app
 
-@Grab('com.xlson.groovycsv:groovycsv:1.3')
 @Grab('com.google.inject:guice:4.2.2')
-import static com.xlson.groovycsv.CsvParser.parseCsv
 import com.google.inject.*
 import java.io.File
 import java.util.ArrayList
-import java.text.SimpleDateFormat
-import static groovy.io.FileType.DIRECTORIES
 import java.util.concurrent.BlockingQueue
 import java.util.concurrent.LinkedBlockingQueue
 
@@ -62,7 +58,7 @@ class MiningFramework {
             
                 printStartAnalysis()                
                 
-                ArrayList<Project> projectList = getProjectList()
+                ArrayList<Project> projectList = Project.getProjectList()
 
                 framework.setProjectList(projectList)
                 framework.start()
@@ -124,50 +120,6 @@ class MiningFramework {
 
     public void setProjectList(ArrayList<Project> projectList) {
         this.projectList = projectList
-    }
-
-    static ArrayList<Project> getProjectList() {
-        ArrayList<Project> projectList = new ArrayList<Project>()
-
-        String projectsFile = new File(getInputPath()).getText()
-        def iterator = parseCsv(projectsFile)
-        for (line in iterator) {
-            String name = line[0]
-            String path = line[1]
-
-            boolean relativePath
-            try {
-                relativePath = line[2].equals("true")
-            } catch(ArrayIndexOutOfBoundsException e) {
-                relativePath = false
-            }
-
-            if(relativePath) 
-                projectList.addAll(getProjects(name, path))
-            else {
-                Project project = new Project(name, path)
-                projectList.add(project)
-            }
-        }
-
-        return projectList
-    }
-
-    static ArrayList<Project> getProjects(String directoryName, String directoryPath) {
-        ArrayList<Project> projectList = new ArrayList<Project>()
-
-        File directory = new File(directoryPath)
-        directory.traverse(type: DIRECTORIES, maxDepth: 0) {
-             
-             // Checking if it's a git project.
-             String filePath = it.toString()
-             if(new File("${filePath}/.git").exists()) {
-                 Project project = new Project("${directoryName}/${it.getName()}", filePath)
-                 projectList.add(project)
-             }
-        }
-
-        return projectList
     }
 
     void setArguments(Arguments arguments) {
