+++ src/main/project/Project.groovy
@@ -2,6 +2,10 @@ package main.project
 
 import main.util.ProcessRunner
 import main.exception.UnexpectedOutputException
+import static groovy.io.FileType.DIRECTORIES
+import static main.app.MiningFramework.arguments
+import static com.xlson.groovycsv.CsvParser.parseCsv
+@Grab('com.xlson.groovycsv:groovycsv:1.3')
 
 class Project {
     
@@ -112,5 +116,49 @@ class Project {
             return [projectOwner, projectName]
         }
         return []
+    }
+
+     static ArrayList<Project> getProjectList() {
+        ArrayList<Project> projectList = new ArrayList<Project>()
+
+        String projectsFile = new File(arguments.getInputPath()).getText()
+        def iterator = parseCsv(projectsFile)
+        for (line in iterator) {
+            String name = line[0]
+            String path = line[1]
+
+            boolean relativePath
+            try {
+                relativePath = line[2].equals("true")
+            } catch(ArrayIndexOutOfBoundsException e) {
+                relativePath = false
+            }
+
+            if(relativePath) 
+                projectList.addAll(getProjects(name, path))
+            else {
+                Project project = new Project(name, path)
+                projectList.add(project)
+            }
+        }
+
+        return projectList
+    }
+
+    static ArrayList<Project> getProjects(String directoryName, String directoryPath) {
+        ArrayList<Project> projectList = new ArrayList<Project>()
+
+        File directory = new File(directoryPath)
+        directory.traverse(type: DIRECTORIES, maxDepth: 0) {
+             
+             // Checking if it's a git project.
+             String filePath = it.toString()
+             if(new File("${filePath}/.git").exists()) {
+                 Project project = new Project("${directoryName}/${it.getName()}", filePath)
+                 projectList.add(project)
+             }
+        }
+
+        return projectList
     }
 }
